+++ pmp-reactjs-ui/public/i18n/ara.json
@@ -17,7 +17,10 @@
         "urlTooLong": "يجب ألا يسمح URI بأكثر من {{length}} حرف.",
         "invalidUrl": "أدخل عنوان URI صالحًاe",
         "duplicateUrl": "غير مسموح بعناوين URI المكررة",
-        "errorInResponse": "أثناء جلب جميع سياسات شركاء المصادقة المعتمدة، واجهنا خطأ"
+        "errorInResponse": "أثناء جلب جميع سياسات شركاء المصادقة المعتمدة، واجهنا خطأ",
+        "filterBtn": "منقي",
+        "resetFilter": "تصفية إعادة تعيين",
+        "itemsPerPage": "مواد لكل صفحة"
     },
     "dashboard": {
         "welcomeMsg": "مرحباً {{firstName}} {{lastName}}",
@@ -131,8 +134,6 @@
         "policies": "سياسات",
         "requestPolicyBtn": "سياسة الطلب",
         "listOfPolicies": "قائمة طلبات السياسة",
-        "filterBtn": "منقي",
-        "resetFilter": "تصفية إعادة تعيين",
         "partnerId": "معرف الشريك",
         "partnerType": "معرف الشريك",
         "policyGroupName": "مجموعة السياسات",
@@ -146,7 +147,6 @@
         "selectPolicyGroup": "حدد مجموعة السياسات",
         "selectPolicyName": "حدد اسم السياسة",
         "selectStatus": "حدد الحالة",
-        "itemsPerPage": "مواد لكل صفحة",
         "errorInPoliciesList": "هناك بعض الخطأ في جلب السياسات. حاول مرة أخرى في وقت لاحق!"
     },
     "viewPolicyDetails": {
@@ -420,7 +420,7 @@
         "PMS_PMP_51": "السياسة المقدمة لا تنتمي إلى مجموعة سياسة الشريك",
         "PMS_PMP_052": "شهادة الشريك غير متوفرة",
         "PMS_PMP_014": "السياسة غير موجودة",
-        "PMS_PMS_022": "تم توفير مفتاح واجهة برمجة تطبيقات الشريك غير نشط.",
+        "PMS_PMS_022": "لقد تم بالفعل إلغاء تنشيط مفتاح API",
         "PMS_PMP_018": "انتهت صلاحية السياسة.",
         "PMS_PRT_108": "مجموعة السياسات إلزامية لـ Auth_Partner",
         "PMS_PRT_110": "الاسم الذي تم إدخاله موجود بالفعل. أدخل اسمًا فريدًا وأرسله.",
@@ -497,6 +497,24 @@
         "PMS_POLICY_ERROR_004": "لم يتم العثور على سياسة المطابقة.",
         "PMS_POLICY_ERROR_005": "معرف مجموعة السياسات فارغ.",
         "PMS_POLICY_ERROR_006": "حدث خطأ أثناء جلب قائمة عملاء OIDC.",
-        "PMS_POLICY_ERROR_007": "حدث خطأ أثناء جلب طلبات مفتاح API"
+        "PMS_POLICY_ERROR_007": "حدث خطأ أثناء جلب طلبات مفتاح API",
+        "PMS_ESI_001": "المفتاح العام للعميل موجود بالفعل.",
+        "PMS_ESI_002": "لم تتم الموافقة على تعيين سياسة الشريك. يرجى التحقق من حالة التعيين.",
+        "PMS_ESI_003": "سياسة لها مصادقة إلزامية. لذلك لا يمكن إنشاء العميل لنفسه.",
+        "PMS_ESI_004": "العميل غير موجود بالمعرف المحدد.",
+        "PMS_ESI_005": "الشريك ليس لديه مطالبات المستخدم",
+        "PMS_ESI_006": "ليس لدى الشريك مرجع سياق المصادقة",
+        "PMS_ESI_007": "فشلت معالجة المفتاح العام",
+        "PMS_ESI_008": "تم إلغاء تنشيط العميل بالفعل.",
+        "PMS_PRT_056": "نظرا لأن langCode غير مدعوم",
+        "PMS_PRT_057": "عدم تطابق نوع السياسة. يمكن استخدام سياسة المصادقة فقط لإنشاء عميل OIDC.",
+        "PMS_PRT_058": "شريك المصادقة غير موجود بهذا المعرف",
+        "PMS_PRT_059": "يمكن لـ Auth_partner فقط إنشاء عميل OIDC. لا يُسمح لأنواع الشركاء الأخرى بإنشاء عميل OIDC.",
+        "PMS_PRT_066": "يجب ألا يحتوي معرف الشريك على أية مسافات",
+        "PMS_PRT_096": "Json غير صالح",
+        "PMS_CONSENT_ERROR_001": "حدث خطأ أثناء جلب موافقة الشريك.",
+        "PMS_CONSENT_ERROR_002": "حدث خطأ أثناء حفظ موافقة الشريك.",
+        "PMP-AUT-020": "استثناء خطأ التحليل",
+        "PMP-AUT-021": "استثناء التدقيق من العميل"
     }
 }
